chore: renaming some store & providers
