--- frontend/core/hooks/useArticles.ts
@@ -1,1 +1,0 @@
-export { default } from '~/stores/articles/hooks'
