--- frontend/core/providers/index.tsx
@@ -1,3 +1,0 @@
-export { default as GlobalLayout } from './GlobalLayout'
-export { default as GraphQLProvider } from './GraphQL'
-export { parseRouteInfo } from './helper'
