+++ frontend/core/providers/ssr.ts
