+++ frontend/main/src/app/[community]/layout.tsx
@@ -1,6 +1,6 @@
 import { LOCALE } from '~/const/i18n'
-import { GlobalLayout } from '~/providers'
-import { getCommunityInfo, getLocaleData } from '~/providers/domain'
+import GlobalProvider from '~/providers/Global'
+import { getCommunityInfo, getLocaleData } from '~/providers/ssr'
 import MainProvider from '~/stores/provider'
 import Client from './Client'
 
@@ -14,9 +14,9 @@ export default async ({ children, params }) => {
 
   return (
     <MainProvider initData={community} locale={LOCALE.EN} localeData={JSON.stringify(localeData)}>
-      <GlobalLayout>
+      <GlobalProvider>
         <Client>{children}</Client>
-      </GlobalLayout>
+      </GlobalProvider>
     </MainProvider>
   )
 }
